src/requirements.txt
python-dotenv
parse
pathlib

#Scrapy Additional Package
Scrapy
scrapy-rotating-proxies

#Mongo
pymongo